fix(terrain): resolve naming conflicts and improve terrain system integration
- Consolidated duplicate TerrainType enums, using existing TerrainSegment version
- Updated terrain factors to match Session 7 research (sand: 2.1, snow: 2.5)
- Removed duplicate TerrainDetectionResult and TerrainDetectionError structs
- Added Sendable conformance and missing properties for compatibility
- Enhanced TerrainSegment with iconName and colorIdentifier properties
- Regenerated Xcode project with xcodegen to include all new files

Technical improvements:
- Unified terrain type system across all components
- Enhanced terrain factor accuracy based on military research
- Improved Swift 6 concurrency compliance
- Better integration between terrain detection and existing data models
