+++ RuckMap/Models/TerrainDetector.swift
@@ -3,6 +3,7 @@ import CoreLocation
 import CoreMotion
 import MapKit
 import Observation
+import SwiftData
 
 /// Errors that can occur during terrain detection
 enum TerrainDetectionError: LocalizedError, Sendable {
@@ -28,75 +29,6 @@ enum TerrainDetectionError: LocalizedError, Sendable {
     }
 }
 
-/// Terrain types for rucking with associated friction coefficients (μ)
-enum TerrainType: String, CaseIterable, Codable, Sendable {
-    case pavedRoad = "paved_road"
-    case trail = "trail"
-    case gravel = "gravel"
-    case sand = "sand"
-    case mud = "mud"
-    case snow = "snow"
-    case stairs = "stairs"
-    case grass = "grass"
-    
-    /// Human-readable display name
-    var displayName: String {
-        switch self {
-        case .pavedRoad: return "Paved Road"
-        case .trail: return "Trail"
-        case .gravel: return "Gravel"
-        case .sand: return "Sand"
-        case .mud: return "Mud"
-        case .snow: return "Snow"
-        case .stairs: return "Stairs"
-        case .grass: return "Grass"
-        }
-    }
-    
-    /// Terrain friction coefficient (μ) for calorie calculations
-    /// Based on biomechanical research for walking/rucking efficiency
-    /// Updated values align with Session 7 research and Pandolf equation requirements
-    var terrainFactor: Double {
-        switch self {
-        case .pavedRoad: return 1.0  // Baseline efficiency (Pandolf standard)
-        case .trail: return 1.2      // 20% increase in energy expenditure
-        case .gravel: return 1.3     // 30% increase for loose surfaces
-        case .sand: return 2.1       // 110% increase due to energy loss (Session 7 research)
-        case .mud: return 1.8        // 80% increase for soft/sticky terrain
-        case .snow: return 2.5       // 150% increase due to poor traction (Session 7: 6" snow)
-        case .stairs: return 2.0     // 100% increase for vertical movement
-        case .grass: return 1.2      // 20% increase similar to trails
-        }
-    }
-    
-    /// SF Symbol icon name for UI representation
-    var iconName: String {
-        switch self {
-        case .pavedRoad: return "road.lanes"
-        case .trail: return "figure.hiking"
-        case .gravel: return "circle.grid.3x3.fill"
-        case .sand: return "beach.umbrella"
-        case .mud: return "cloud.rain"
-        case .snow: return "snowflake"
-        case .stairs: return "stairs"
-        case .grass: return "leaf"
-        }
-    }
-    
-    /// Army green design system color identifier
-    var colorIdentifier: String {
-        switch self {
-        case .pavedRoad: return "armyGreen.secondary"
-        case .trail: return "armyGreen.primary"
-        case .gravel: return "armyGreen.tertiary"
-        case .sand: return "armyGreen.accent"
-        case .mud: return "armyGreen.dark"
-        case .snow: return "armyGreen.light"
-        case .stairs: return "armyGreen.bright"
-        case .grass: return "armyGreen.natural"
-        }
-    }
-}
 
 /// Terrain detection result with confidence metrics
 struct TerrainDetectionResult: Sendable {
@@ -104,17 +36,25 @@ struct TerrainDetectionResult: Sendable {
     let confidence: Double // 0.0 to 1.0
     let timestamp: Date
     let detectionMethod: DetectionMethod
+    let isManualOverride: Bool
     
     enum DetectionMethod: Sendable {
         case motion
         case location
         case mapKit
+        case mapKitOnly
+        case motionOnly
         case fusion
         case manual
+        case manualOverride
     }
     
     var isHighConfidence: Bool {
         confidence >= 0.85
+    }
+    
+    var terrainFactor: Double {
+        terrainType.terrainFactor
     }
 }
